refactor : remove event, history and move to directory
