+++ java-module/mc-o11y-manager/src/main/java/com/mcmp/o11ymanager/entity/TargetEntity.java
@@ -46,7 +46,8 @@ public class TargetEntity {
   private String csp;
 
   @Enumerated(EnumType.STRING)
-  private TargetStatus targetStatus;
+  private TargetStatus status;
+
 
   @Enumerated(EnumType.STRING)
   private TargetAgentTaskStatus monitoringAgentTaskStatus;
