--- java-module/mc-o11y-manager/src/main/java/com/mcmp/o11ymanager/event/BaseDomainEvent.java
@@ -1,22 +1,0 @@
-package com.mcmp.o11ymanager.event;
-
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-@Getter
-public abstract class BaseDomainEvent implements DomainEvent {
-    private final String eventId;
-    private final LocalDateTime occurredOn;
-    
-    protected BaseDomainEvent() {
-        this.eventId = UUID.randomUUID().toString();
-        this.occurredOn = LocalDateTime.now();
-    }
-
-    @Override
-    public String getEventType() {
-        return this.getClass().getSimpleName();
-    }
-}
