--- java-module/mc-o11y-manager/src/main/java/com/mcmp/o11ymanager/event/DomainEvent.java
@@ -1,9 +1,0 @@
-package com.mcmp.o11ymanager.event;
-
-import java.time.LocalDateTime;
-
-public interface DomainEvent {
-    String getEventId();
-    LocalDateTime getOccurredOn();
-    String getEventType();
-}
