--- java-module/mc-o11y-manager/src/main/java/com/mcmp/o11ymanager/event/EventPublisher.java
@@ -1,5 +1,0 @@
-package com.mcmp.o11ymanager.event;
-
-public interface EventPublisher {
-    void publish(DomainEvent event);
-}
