+++ java-module/mc-o11y-manager/src/main/java/com/mcmp/o11ymanager/facade/FluentBitFacadeService.java
@@ -6,13 +6,7 @@ import com.mcmp.o11ymanager.enums.Agent;
 import com.mcmp.o11ymanager.enums.AgentAction;
 import com.mcmp.o11ymanager.enums.ResponseStatus;
 import com.mcmp.o11ymanager.enums.SemaphoreInstallMethod;
-import com.mcmp.o11ymanager.event.AgentHistoryEvent;
-import com.mcmp.o11ymanager.event.AgentHistoryFailEvent;
-import com.mcmp.o11ymanager.exception.config.FileReadingException;
 import com.mcmp.o11ymanager.global.aspect.request.RequestInfo;
-import com.mcmp.o11ymanager.infrastructure.util.ChaCha20Poly3105Util;
-import com.mcmp.o11ymanager.infrastructure.util.CheckUtil;
-import com.mcmp.o11ymanager.model.agentHealth.SshConnection;
 import com.mcmp.o11ymanager.model.host.TargetAgentTaskStatus;
 import com.mcmp.o11ymanager.model.semaphore.Task;
 import com.mcmp.o11ymanager.oldService.domain.interfaces.SshService;
@@ -39,8 +33,6 @@ public class FluentBitFacadeService {
   private static final Lock agentTaskStatusLock = new ReentrantLock();
   private final RequestInfo requestInfo;
   private final SemaphoreDomainService semaphoreDomainService;
-  private final ApplicationEventPublisher event;
-  private final SshService sshService;
   private final SchedulerFacadeService schedulerFacadeService;
   private final FluentBitConfigFacadeService fluentBitConfigFacadeService;
 
@@ -63,18 +55,6 @@ public class FluentBitFacadeService {
     // 5. task ID, task status 업데이트
     targetService.updateLogAgentTaskStatusAndTaskId(nsId, mciId, targetId, TargetAgentTaskStatus.INSTALLING,
         String.valueOf(task.getId()));
-
-    // 6. 이력 남기기
-    AgentHistoryEvent successEvent = AgentHistoryEvent.builder()
-        .requestId(requestInfo.getRequestId())
-        .nsId(nsId)
-        .mciId(mciId)
-        .targetId(targetId)
-        .agentAction(AgentAction.LOG_AGENT_INSTALL_STARTED)
-        .reason("")
-        .build();
-
-    event.publishEvent(successEvent);
 
     // 7. 스케줄러 등록
     schedulerFacadeService.scheduleTaskStatusCheck(requestInfo.getRequestId(), task.getId(), nsId, mciId, targetId,
@@ -99,17 +79,6 @@ public class FluentBitFacadeService {
     targetService.updateLogAgentTaskStatusAndTaskId(nsId, mciId, targetId, TargetAgentTaskStatus.UPDATING,
             String.valueOf(task.getId()));
 
-    // 5. 이력 남기기
-    AgentHistoryEvent successEvent = AgentHistoryEvent.builder()
-            .requestId(requestInfo.getRequestId())
-            .nsId(nsId)
-            .mciId(mciId)
-            .targetId(targetId)
-            .agentAction(AgentAction.LOG_AGENT_UPDATE_STARTED)
-            .reason("")
-            .build();
-
-    event.publishEvent(successEvent);
 
     // 6. 스케줄러 등록
     schedulerFacadeService.scheduleTaskStatusCheck(requestInfo.getRequestId(), task.getId(), nsId, mciId, targetId,
@@ -133,17 +102,6 @@ public class FluentBitFacadeService {
     targetService.updateLogAgentTaskStatusAndTaskId(nsId, mciId, targetId, TargetAgentTaskStatus.UNINSTALLING,
         String.valueOf(task.getId()));
 
-    // 5) 이력 남기기
-    AgentHistoryEvent successEvent = AgentHistoryEvent.builder()
-        .requestId(requestInfo.getRequestId())
-        .nsId(nsId)
-        .mciId(mciId)
-        .targetId(targetId)
-        .agentAction(AgentAction.LOG_AGENT_UNINSTALL_STARTED)
-        .reason("")
-        .build();
-
-    event.publishEvent(successEvent);
     // 6) 스케줄러 등록
     schedulerFacadeService.scheduleTaskStatusCheck(requestInfo.getRequestId(), task.getId(), nsId, mciId, targetId,
         SemaphoreInstallMethod.UNINSTALL, Agent.FLUENT_BIT);
@@ -178,17 +136,6 @@ public class FluentBitFacadeService {
 
       agentTaskStatusLock.unlock();
 
-      AgentHistoryEvent successEvent = AgentHistoryEvent.builder()
-              .requestId(requestInfo.getRequestId())
-              .nsId(nsId)
-              .mciId(mciId)
-              .targetId(targetId)
-              .agentAction(AgentAction.ENABLE_FLUENT_BIT)
-              .reason("")
-              .build();
-
-      event.publishEvent(successEvent);
-
       results.add(ResultDTO.builder()
               .nsId(nsId)
               .mciId(mciId)
@@ -201,16 +148,6 @@ public class FluentBitFacadeService {
 
       targetService.updateLogAgentTaskStatus(nsId, mciId, targetId, TargetAgentTaskStatus.IDLE);
 
-      AgentHistoryFailEvent failEvent = AgentHistoryFailEvent.builder()
-              .requestId(requestInfo.getRequestId())
-              .reason(e.getMessage())
-              .agentAction(AgentAction.ENABLE_FLUENT_BIT)
-              .nsId(nsId)
-              .mciId(mciId)
-              .targetId(targetId)
-              .build();
-
-      event.publishEvent(failEvent);
 
       results.add(ResultDTO.builder()
               .nsId(nsId)
@@ -250,15 +187,6 @@ public class FluentBitFacadeService {
       // 4. 작업 완료 후 idle 변경
       targetService.updateLogAgentTaskStatus(nsId, mciId, targetId, TargetAgentTaskStatus.IDLE);
 
-      AgentHistoryEvent successEvent = AgentHistoryEvent.builder()
-              .requestId(requestInfo.getRequestId())
-              .nsId(nsId)
-              .mciId(mciId)
-              .targetId(targetId)
-              .agentAction(AgentAction.DISABLE_FLUENT_BIT)
-              .reason("")
-              .build();
-      event.publishEvent(successEvent);
 
       results.add(ResultDTO.builder()
               .nsId(nsId)
@@ -273,16 +201,6 @@ public class FluentBitFacadeService {
       agentTaskStatusLock.unlock();
 
       targetService.updateLogAgentTaskStatus(nsId, mciId, targetId, TargetAgentTaskStatus.IDLE);
-
-      AgentHistoryFailEvent failEvent = AgentHistoryFailEvent.builder()
-              .requestId(requestInfo.getRequestId())
-              .reason(e.getMessage())
-              .agentAction(AgentAction.DISABLE_FLUENT_BIT)
-              .nsId(nsId)
-              .mciId(mciId)
-              .targetId(targetId)
-              .build();
-      event.publishEvent(failEvent);
 
       results.add(ResultDTO.builder()
               .nsId(nsId)
@@ -322,15 +240,6 @@ public class FluentBitFacadeService {
 
       targetService.updateLogAgentTaskStatus(nsId, mciId, targetId, TargetAgentTaskStatus.IDLE);
 
-      AgentHistoryEvent successEvent = AgentHistoryEvent.builder()
-              .requestId(requestInfo.getRequestId())
-              .nsId(nsId)
-              .mciId(mciId)
-              .targetId(targetId)
-              .reason("")
-              .build();
-
-      event.publishEvent(successEvent);
 
       results.add(ResultDTO.builder()
               .nsId(nsId)
@@ -343,17 +252,6 @@ public class FluentBitFacadeService {
       agentTaskStatusLock.unlock();
 
       targetService.updateLogAgentTaskStatus(nsId, mciId, targetId, TargetAgentTaskStatus.IDLE);
-
-      AgentHistoryFailEvent failEvent = AgentHistoryFailEvent.builder()
-              .requestId(requestInfo.getRequestId())
-              .reason(e.getMessage())
-              .agentAction(AgentAction.ENABLE_FLUENT_BIT)
-              .nsId(nsId)
-              .mciId(mciId)
-              .targetId(targetId)
-              .build();
-
-      event.publishEvent(failEvent);
 
       results.add(ResultDTO.builder()
               .nsId(nsId)
