--- java-module/mc-o11y-manager/src/main/java/com/mcmp/o11ymanager/infrastructure/event/SpringEventPublisher.java
@@ -1,18 +1,0 @@
-package com.mcmp.o11ymanager.infrastructure.event;
-
-import com.mcmp.o11ymanager.event.DomainEvent;
-import com.mcmp.o11ymanager.event.EventPublisher;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class SpringEventPublisher implements EventPublisher {
-    private final ApplicationEventPublisher applicationEventPublisher;
-
-    @Override
-    public void publish(DomainEvent event) {
-        applicationEventPublisher.publishEvent(event);
-    }
-}
